... and use correct types.

+++ subprocess/subprocess_windows.go
@@ -391,7 +391,7 @@ func (sub *Subprocess) BottomHalf(d *SubprocessData, sig chan *SubprocessResult)
 	var runState runningState
 
 	for result.SuccessCode == 0 && waitResult == syscall.WAIT_TIMEOUT {
-		waitResult, _ = syscall.WaitForSingleObject(hProcess, sub.TimeQuantum.Nanoseconds() / 1000000)
+		waitResult, _ = syscall.WaitForSingleObject(hProcess, uint32(sub.TimeQuantum.Nanoseconds() / 1000000))
 		if waitResult != syscall.WAIT_TIMEOUT {
 			break
 		}
